Easier introduction of new damage types.
rP20203 introduced a DamageTypes.js global script similar to the resources one. However, we never actually need to refer to this script since we can always use the damage types provided by the template/context/object we are looping over/...
There is one exception to this for AI weighting of damage types. However, since damage types are not stored in files, this is strictly equivalent to hardcoding them in the global script and was deemed acceptable.

Patch By: freagarach

Reviewed By: wraitii

Refs #4801 (by invalidating it for now, though such helper files might be useful in the future if damage types require more metadata).

Differential Revision: https://code.wildfiregames.com/D1938


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22527 3db68df2-c116-0410-a063-a993310a9797