Add virtual CreateJSObject to IGUIObject, split from GetJSObject, so that inheriting classes can extend the JS object upon construction, refs #5442, D2136.
Remove the comment from rP5154 that allures the reader to believe there is a memory leak which is not and has not been the case, refs D1700.
Remove the useless JSAutoRequest from GetJSObject if the object was constructed already (performance improvement), following cx assignment in rP14496, following JSAutoRequest addition in rP14877 (that also had removed other useles JSAutoRequest calls but not this one).
Don't change JS::PersistentRooted to JS::Heap until SpiderMonkey is updated, refs D1700.
Provide access to GUI, ScriptHandlers and JSObject for inheriting classes.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22593 3db68df2-c116-0410-a063-a993310a9797