Fix unused glooxwrapper variable following rP19703, refs #2305, #5294, #5550.
The class didn't leak the m_Wrapped gloox Jingle for this program since it's not set to owned, and it's not set to owned because gloox deletes it.
It would have leaked if some other app would have used glooxwrapper with owned = true, if gloox has a situation for that.

Differential Revision: https://code.wildfiregames.com/D2090



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22764 3db68df2-c116-0410-a063-a993310a9797