Add a system component to handle stat modifiers, make technologies and auras use this common interface.
The ModifiersManager system component provides an interface to add and remove modifiers, and get modified stats.
The goal is to merge all the different stat-modifying systems 0 A.D. has implemented over the years.
This commit makes technologies and auras use ModifiersManager. Some cheats and AI bonuses also have a similar stat-modifying effect that have not yet been updated.

Further, this system component makes it possible for e.g. triggers to easily add modifiers, enabling the writing of Castle Blood Automatic, RPG or Tower Defense maps without the need for mods or hacks.

The 'Modifier' name was preferred over 'Modification' as it is shorter and more readable, along with the logic that 'modifiers' store 'modifications' and this stores modifiers. Renaming of other functions and classes has been left for future work for now.

Internally, this uses a JS data structure. If performance issues arise with it in the future, this data structure or the whole component could be moved to C++.
The performance has been tested to be about as fast as the current implementations (and specifically much faster for global auras with no icons). Testing showed that sending value modification messages was by far the slowest part.

Comments by: leper, Stan, elexis

Differential Revision: https://code.wildfiregames.com/D274


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22767 3db68df2-c116-0410-a063-a993310a9797