Unify loadgame and savegame dialog following rP13579, refs #2030.
This way the save dialog gains the sortable list and details section feature from rP19351 / D246 by removing a folder of code, refs #4413.
Fixes a style warning about the old dialog using a style for an 'olist' type for a 'list' type, refs rP22792 / D2232.

Subsequently move the gui/common/ helper file to the page code, meaning it's not loaded for every GUI page needlessly anymore.
Fix map preview aspect ratio.
Use object-oriented code for the moved savegame code, refs #5387, but conserve the rest of the code to limit the changes for auditability.

Differential Revision: https://code.wildfiregames.com/D2290
Tested on: Jenkins



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22923 3db68df2-c116-0410-a063-a993310a9797