Delete dead "chat" simulation command, "chat" GUIInterface notification type, "aichat" NetClient message and useless "debug-print" simulation command.
The "chat" simulation command and according GUIInterface notification became unused which translated simulation chat messages ("aichat") in rP14991.
That commit also introduced the "aichat" network message parsing code that is never triggered since the NetClient never creates messages of that type.
The "debug-print" command introduced by rP7653 sends a GUI value to the simulation and prints it to the log, which is already covered by the JS console since rP469.

Differential Revision: https://code.wildfiregames.com/D2407



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23137 3db68df2-c116-0410-a063-a993310a9797