Do not face point after movement for formation walk
After formation member ends walking, he tries to face target point. However that eliminates rotation set by SetInPosition from formation component. Partly fixed in rP23248 with setting correct rotation in idle state, but one could still notice that there is glitch when unit does quick turn to target point and then back. However that solution would be still required if logic in SetInPosition would not be changed to always set correct rotation. Another alternative would be to unset unit at position with every walking order but that would have the same effect and this is just more performance friendly to not splice member from position every time.

Notice that information from inPosition is currently not used, so that change in it is not braking anything.


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23346 3db68df2-c116-0410-a063-a993310a9797