Update quotes.txt
Removed this one: “The greatest glory for women is to be least talked about by men, whether for good or ill.” \n— Pericles in his Funeral Oration for Athenians that died in the first year of the war (Thucydides, “History of the Peloponnesian War”, II. 45.2)

Not a vital quote. The project can survive without it.