Removes unused gfx::GetMonitorSize function
GetMonitorSize was added in rP216. Its usage wasn't added. We don't need to have the custom implementation of the function, since SDL2 can do this. Also it doesn't make sense to get the default monitor size in multi-monitor environments since the application window might be located on each of them.

Patch By: linkmauve

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23426 3db68df2-c116-0410-a063-a993310a9797