React to target death properly when chasing
Differential Revision: https://code.wildfiregames.com/D2517
Reviewed by: Freagarach
Comments by: elexis, Stan
Fixes: #5610 #5656

Introduce likelyFailure check in the movementupdate for chasing state missing since rP22526.
Use new order and substate for Chasing to get into the wanted range and avoid that way issue with running - walking tracking or animation variants.

Fixing issue when attacker would not react to the death of the target while chasing it in stances where results for abandoning chasing are always false and also do not rely on timer ending chasing for the rest of stances.

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23566 3db68df2-c116-0410-a063-a993310a9797