Revert rP23505 / r23505
Summary: Having a mouse event and a hotkey for the same hardware key turned out to be bad. If the mouse is above a gui element that handles the middle mouse event and the middle mouse is released, this event will be consumed by the gui handler of this gui element. Therefore the hotkey handler does not know that the key was released. If we change that behaviour typing in input fields would trigger hotkeys. So this seems unsolvable without adding special handling for the middle mouse case, which would be ugly.


Fixes: #5695

Differential Revision: https://code.wildfiregames.com/D2700




git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23589 3db68df2-c116-0410-a063-a993310a9797