Remove some variables from serialisation and update this.animations name to reflect what it actually holds [Formations.js] Needs ReviewPublic
Differential Revision: https://code.wildfiregames.com/D2707
Comments by: elexis, Stan, wraitii, bb

Remove variables were not needed to be serialised or stored inside component
Create `variablesToSerialize` to hold list of variables, that need to be Serialized.
Rename `this.template.Animations` to `this.template.AnimationVariants` (and anything related) as it does not hold animation names but names of animation variants what is misleading.
Do not create timer when deserializing as that would cause oos and timer already exists.
this.Init has to be called from Deserialize because it is not called by default

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23694 3db68df2-c116-0410-a063-a993310a9797