Trigger an exit-reentry when the target entity of an order is renamed
This lets sunit AI FSM states correctly handle target entity renaming by processing a message when that happens. The default behaviour is to leave-reenter the state, which re-runs sanity checks and optionally picks a better behaviour.

UnitMotion is still not made aware of entity renaming, as the leave-enter makes it irrelevant in practice. It still may be a good idea to implement that someday.

Fixes the concern raised at rP20393.
Fixes #5584

Comments by: bb, Freagarach

Reported by: minohaka, bb, Freagarach

Differential Revision: https://code.wildfiregames.com/D2735



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23708 3db68df2-c116-0410-a063-a993310a9797