Tweak wonder/pyramid cost and remove healing aura from wonders.
Wonders cost food previously, which is not in line with their look or other buildings.
Their healing aura is removed to make the slightly easier to capture.

Kushites pyramids are adjusted slightly, and the monumental architecture tech moved to the CC to match the Persian.

Patch by: Nescio

Accepted By: borg-, Feldfeld

Differential Revision: https://code.wildfiregames.com/D2686



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23824 3db68df2-c116-0410-a063-a993310a9797