Add an interface for Reinforcement Learning.
Implement a simple HTTP server to start games, receive the gamestate and pass commands to the simulation.
This is mainly intended for training reinforcement learning agents in 0 AD. As such, a python client and a small example are included.

This option can be enabled using the -rl-interface flag.

Patch by: irishninja

Reviewed By: wraitii, Itms

Fixes #5548

Differential Revision: https://code.wildfiregames.com/D2199



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23917 3db68df2-c116-0410-a063-a993310a9797