Fix a compartment mismatch in XmppClient, causing crashes in MP games.
Added in rP22856, which changed lobby GUI messages to JS::Values, requiring a real context. The original code mistakenly inverted the owning script interfaces.

Given the reproducibility discovered in SM52, the timeline of the bug, and the nature of the issues encountered in MP, this is a rather safe fix for #5655.

Reviewed By: Itms

Fixes #5655

Differential Revision: https://code.wildfiregames.com/D2922



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24116 3db68df2-c116-0410-a063-a993310a9797