Fix UpdateComponents logic for pathfinding following rP22902
Following rP22902, paths requested at turn N were set-up to be computed between the end of turn N and the start of turn N+1 (which would ultimately allow threading this computation), via calls to 'StartProcessingMoves' and 'FetchAsyncResultsAndSendMessages'.

However, the call to UpdateGrid() remained at the start of turn N+1, between the 'start' and 'fetch' calls. Since all paths are currently computed on the 'start' call, this means all paths are computed on a (possibly) dirty pathfinder grid.
In particular, this leads to OOS on rejoin since the rejoiner will recompute the grid before computing the outstanding paths.

This would also obviously be buggy in a threaded environment, since some paths might be computed on the fresh and some on the dirty grid.

Finally, MT_TurnStart was sent before the paths were computed, which might lead to further pathfinder grid changes (not a crashing problem without threading, but still conceptually odd). The 'fetch' call is thus moved before it.
This thus fixes rP22902/D1918, after rP22979 already fixed a first issue.

Since the grid is now only updated at the end of a turn, we need to ensure that it is correct on Turn 0, thus the pathfinder recomputes it on InitGame.

Refs D14

Reported by: Itms

Fixes #5851

Differential Revision: https://code.wildfiregames.com/D3064



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24142 3db68df2-c116-0410-a063-a993310a9797