Remove serializablePrototype code
The script interface has code to serialize/deserialize JS objects with a user-defined prototype. That code is usable in the AI, but currently unused (and there are no plans to use it in the long run).

Removing it allows removing more code down the line, which helps with the SM45-52 migration.

Collaboration with itms.

Refs #4893

Differential Revision: https://code.wildfiregames.com/D2897



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24157 3db68df2-c116-0410-a063-a993310a9797