Mass rename CxPrivate to CmptPrivate.
As part of the SM45->52 migration, a ScriptInterface becomes a wrapper around a JSCompartment, not a JSContext, thus we ought to store private data for the compartment and not the context.
This is a mass rename of CxPrivate to CmptPrivate to match that before the actual changes.

Part of the SM52 migration, stage: SM45 compatible

Patch by: Itms

Tested By: Freagarach

Refs #4893

Differential Revision: https://code.wildfiregames.com/D3089



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24177 3db68df2-c116-0410-a063-a993310a9797