Fix JS OOM in tests following rP24180
rP24180 made ScriptInterface not recreate a new JSContext since it becomes a compartment wrapper, but that means we need to GC tests or they might OOM.

To make it mostly seamless, GC on any test setup. The JS tests are pretty close to the 16Mb limit as it stands so GC them manually too, for good measure.

Refs #4893

Differential Revision: https://code.wildfiregames.com/D3101



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24182 3db68df2-c116-0410-a063-a993310a9797