[SM52 1/2] Upgrade Spidermonkey build system and binaries to 52.9.1
Of note is the change to static linking on *nix (MacOS already used it). The logic is that we can only use one single version of SM, so that advantage of dynamic linking is lost regardless. We might also see performance gains in the future with LTO enabled. It is also slightly easier to distribute the program as a result. Expect a negligible size increase in the binary size (offset somewhat by no longer needing to distribute .so files). Finally, it streamlines the build script.

Also noteworthy in this commit:
- The MacOS build script is folded back into the general build script.
- the perl/sed command is replaced by patching the configuration file, which at least warns if it starts failing in the future.

Binaries for windows provided by @Itms
The bulk of the patching was also done by @Itms.

Tested by: Stan, Freagarach

Refs #4893

Differential Revision: https://code.wildfiregames.com/D3094



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24202 3db68df2-c116-0410-a063-a993310a9797