Deactivate RTTI (except in AtlasUI) on all platforms.
D2768 inherits from a Spidermonkey class, and SM is built without rtti. This won't compile unless we also compile without RTTI, and that diff is required for SM60.
Further, windows was already compiling without RTTI and our coding conventions say to not use it, so this is more of a regularization.

AtlasUI is left unaffected as WXWidgets uses it (and it's a dynamic library so separate enough to do that).

Differential Revision: https://code.wildfiregames.com/D3104



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24214 3db68df2-c116-0410-a063-a993310a9797